backend/requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
Django==5.2.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt[crypto]==5.3.0
django-filter==24.3
dotenv==0.9.9
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
redis==6.4.0
sqlparse==0.5.3
typing_extensions==4.15.0
google-cloud-storage==2.10.0
Pillow==10.0.1
